+++ sound-mind-studio/tests/test_canvas_widget.h
@@ -34,6 +34,7 @@ private slots:
     void mouseReleaseInPickModeEmitsPickStrokeEndedAndEndsTheGesture();
     void changingToolModeAwayFromPickCancelsAnyActiveGesture();
     void setPickSelectionBoundsDrawsAHighlight();
+    void setPickSelectionBoundsIsHiddenWhileALivePreviewIsShowing();
     void setPickSelectionBoundsWithNoValueDrawsNothing();
     void mousePressInSelectModeEmitsSelectStrokeStartedWithAConvertedPoint();
     void mouseMoveAfterPressInSelectModeEmitsSelectStrokeContinued();