+++ sound-mind-studio/tests/test_main_window.h
@@ -177,6 +177,7 @@ private slots:
     void finishPathWithNoNodesPlacedIsANoOp();
     void settingANewProjectResetsPathModeToOff();
     void pastedContentIsPickableAndMovable();
+    void pasteSwitchesToPickModeAndSelectsThePastedRegionImmediately();
     void modifyingAPaintedStrokeAfterCuttingOverItKeepsTheCutRegionSilenced();
     void bringPickedObjectToFrontMovesItAboveLaterStrokesOnTheSameLayer();
     void editingAPickedStrokesPathMovesANodeAndCommitsOnApply();